requirements.txt
fastapi==0.104.1
reportlab==4.2.5
uvicorn[standard]==0.24.0
tensorflow==2.15.0
pillow==10.1.0
numpy==1.26.2
python-multipart==0.0.6
groq==0.4.1
jinja2==3.1.2
opencv-python==4.8.1.78